@@ -1,7 +1,7 @@
 package com.example.app.presentation.feature.menulist
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.example.app.domain.usecase.GetMenusUseCase
 
 class ViewModelFactory(private val getMenusUseCase: GetMenusUseCase = GetMenusUseCase()) : ViewModelProvider.Factory {
